+++ pkg/providers/clientlists/testData/TestResActivation/activation_invalid_version.tf
@@ -4,9 +4,9 @@ provider "akamai" {
 
 resource "akamai_clientlist_activation" "activation_ASN_LIST_1" {
   list_id                 = "12_AB"
-  version                 = 3
+  version                 = 99
   network                 = "STAGING"
-  comments                = "Activation Comments Updated"
-  notification_recipients = ["update_user@example.com"]
-  siebel_ticket_id        = "UPDATED-12345"
+  comments                = "Activation Comments"
+  notification_recipients = ["user@example.com"]
+  siebel_ticket_id        = "ABC-12345"
 }